requirements.txt
pyyaml
python-digitalocean
paramiko
huepy
discord-webhook